Installing eslint (#3)
Install eslint and config for drone CI

Sets up a basic linter using eslint. Run `yarn lint` with an optional `--fix` flag if you want it to automatically correct the errors.

This is is also adds an eslint step to drone.ci to fail any builds that need linting. Because eslint is a development dependency, we need to explicitly call `yarn install --development` prior to running the test.