Indexing Service
Defines an indexing service that will build a document and index it into
Solr given an arbitrary schema. The only requirement is that the
resource being indexed into Solr respond to a `uuid` method that
corresponds to a globally unique identifier that will be used in the
Solr index.

This also refactors the blacklight.yml configuration to use the
environment variables to determine the correct Solr url.