Removing profile option
The profile option was added to our config back when we were having
really slow test problems. Now that it isn't an issue anymore, the tag
is generating a lot of extra output each time tests are run. This
creates unnecessary background "noise" so let's turn it off and add it
when we need to really profile the test suite.