DefaultSchema for resources
Using the attribute_types of our ActiveRecord resouces, we can create a
basic schema to index them into Solr. Commiting can be selectivly
enabled to better performance.

By default, strings are multivalued in Solr because it isn't possible to
determine if a particular attribute in json is multivalued or not. This
could be fixed later.