Set selenium_chrome_headless as default JS driver
Capybara has the notion of a javascript driver, and it's possible to set
the default one to whatever you want. By registering our
selenium_chrome_headless driver as the default one in development mode,
then you can tag your specs with
  `js: true`
as opposed to
  `with_driver: :selenium_chrome_headless`
which is a heck of a lot easier to remember

In CI mode, we define a special driver needed by Drone to execute
properly in that environment. No specs need to change, it's swapped out
in the config based on the environment.

Co-authored-by: Dann Bohn <djb44@psu.edu>