Database cleaner uses transactions except in JS specs
The transaction cleaner strategy is _much_ faster than the trunction strategy. The catch is that the transaction strategy does not work with browser-based feature testing (e.g. selenium, chromedriver, poltergeist). This commit uses the faster transaction-based strategy most of the time, unless the `:js` flag is added to the spec, which in this case will both trigger the chromedriver capybara driver and the slower but necessary truncation strategy.

Benchmarks:
# Before modification
$ bin/rspec
Finished in 1 minute 9.41 seconds

$ bin/rspec spec/models/work_version_spec.rb
Finished in 8.64 seconds

# After modification:
$ bin/rspec
Finished in 51.57 seconds

$ bin/rspec spec/models/work_version_spec.rb
Finished in 3.72 seconds