Upgrade eslint and use babel-eslint parser
eslint was failing due to Stimulus.js [requiring](https://stimulusjs.org/reference/targets#definitions) the use of a `static` keyword in its controllers. This static keyword was not supported by the existing parser that we'd configured eslint to use. After much stackoverflowing, I was able to swap the parser out for a babel one, which works. I believe this is totally legitimate, too, since our webpack is configured to use babel when in production mode, to transpile the ES6 features that we/stimulus use.