Feature test uses abiguous search for a link
The catalog feature test was failing at random times. What appears to be
the problem was finding the link to click on in the search results.
Instead of searching for the link, we'll visit the page directly.

This also includes an added feature to setting up Faker. We can specify
an specific seed in case the randomized results of the Faker gem are
affecting test results. After each test run, Faker's own randomized seed
will be reported. If we want to re-run the same test using the exact
same random output, we can provide the seed as follows:

    FAKER_SEED=1234 bundle exec rspec test_spec.rb