Sync the user's group membership into the application
A list of groups that the user belongs to will be coming out of our auth
proxy. We take this list and sync it into the Groups table which
maintains a relationship with the user's record.

With the user's group membership maintained locally, we can then use the
groups to contruct access controls for enforcing group-based access on
resources.

Co-authored-by: Ryan Schenk <rschenk@gmail.com>
Co-authored-by: Adam Wead <amsterdamos@gmail.com>