Define permissions based on access controls
Creates a set of "levels" for defining access rights to work resources.
An agent, either a user or group, is granted a level of access that
affords them a set of abilities with a given resource.  These levels
include: discover access, for reading metadata but not downloading
content; read access, for both metdata and binary content access; and,
edit access, for all other rights, including the ability to create,
edit, and delete content, as well as create new versions of works.

Two application-wide groups are provided for granting open access, and
access to anyone at Penn State. A Permissions module manages the agents
who are granted different levels of access and ensures that each
combination of agent and level is unique to a particular resource.