Using grant/revoke syntax for permissions
Updates PermissionsBuilder and Permssions to add or remove specific
access controls via a "grant" or "revoke" syntax. This maps to other
systems of permissions such as SQL.

The methods now accept multiple agents to avoid having to call grant or
revoke more than once for the same level of access.