Define visibility for works
Creates a visibility property that is determined using existing
permissions. The visibility may be "set" on the work by granting or
revoking a specific set of access controls. Before publishing, a work is
validated so that its visibility may not be private, and must either be
open access or accessible to Penn State users.