Don't cache group agents in test
The test environment uses DatabaseCleaner and transactional fixtures to
speed things up. Unfortunately, this causes problems when caching
specific database records for our public and authorized group agents.

To get around the problem, we just don't cache these in test. It's not
the most elegant solution, but it's the simplest and requires the least
amount of change.