Use Warden for non-JS feature tests
By-passing OAuth for feature tests that do not require Javascript helps
performance and avoids problems when accessing pages that don't have an
explicit user authentication, such as the catalog page.

We're adding a separate feature test for OAuth as a "sanity check" and
OAuth is still exercised in a Javascript context because the test setup
is multithreaded.