Rest API for publishing new works
Creates a base controller that we can use for any request to the REST
API. Currently, no authentication is done and forgery protection is
ignored. Once we decide how we want to manage authentication, we can
apply the appropriate mechanism.

Adds an ingest endpoint for creating and publishing new works. This will
be used largely for migration.

Additional features include:

* PublishNewWork service for complete end-to-end work publishing
* UpdatePermissionsService for updating the permissions on a work with a
  given hash of agent names
* N.B!! The base controller for the rest endpoint uses NO
authentication!  This will be changed in subsequent commits.
* Incorporates the scholarpshere-client gem for uploading files to S3