Determine access using a user's groups
Fixes a bug where a user who is a member of a group that is granted a
level of access, must be allowed to access that resource. For example,
given a group that is granted read access to a resource, any user that
is a member of that group must also be allowed read access.

One of the issues created with this fix is that the PermissionsBuilder
class must know that "groups" is a means of finding associated agents
from another agent. This blurs the abstract nature of agents a bit, but
would make good fodder for refactoring.