Authorized and open access only applies to groups
If a user is granted read level access, it will cause NoMethodError
because we were not checking the agent type. Open and authorized access
only applies to groups with read level permissions.

This also refactors how we determine public and authorized agents by
defining them in the access control resource.