Add setter methods for permissions
Adds setter methods for the different permission types. We can pass an
array of agent resources to a work and the permissions will be granted
or revoked as needed.

Special consideration is given to read groups and edit users so that
groups defining visibility are not removed unintentionally, or that an
unneeded access control is created for the depositor.