Move aliased creators from works to work versions
Assigns aliased creators to work versions, instead of works. This also revamps the data model, collapsing it, and making the WorkVersionCreation contain the alias to the creator instead of having that in a separate table. BuildNewWorkVersion is also updated to copy existing aliases and their creators to the new version.