Install cocoon gem for easy nested forms
Note that the base `cocoon` gem contains a javascript file that is intended to integrate with asset pipeline. Since we are using webpacker instead of asset pipeline, we must also install the npm/yarn package `cocoon-js`. This package contains the exact same javascript code as the original cocoon gem, but makes it available to the webpack