FileVersionMembershipPolicy
Creates a policy for governing the downloading of files. The policy
targets the file version membership, which is the relation between the
file itself and the work version it is a part of, but the "meat" of the
policy really centers on the work, and not the work version, since
permissions are defined on the work.

The reasoning behind setting the policy on the membership is that we
will need to determine download rights in a number of different contexts
such as within a user's dashboard, or the public application. The policy
should therefore have no restrictions of scope and determine if a file
can be download within any given context.