Display file list based on download policy
The listing of files in both the public-facing and dashboard work view
is now determined based on whether or not the user has download access
to the files.

Because the work version is the object that determines whether or not
they have this ability, we shifted the policy decision from the file
version to the work version.

In order to make this easier to manage, we introduce some  inheritance
between WorkVersionPolicy and Dashboard::WorkVersionPolicy so that the
both policies may determine if the given resource's files can be
downloaded.