Person authority
Using Questioning Authority, we create a Person authority that can be
used to find available creators for a work. The authority searches our
own database for creator records as well as Penn State's identity
management API.

For a given search string, a set of results are returned as a hash
containing given name, surname, and additional relevant information such
as the user's Penn State ID, email, and OrcID, if available.

Included in this commit is the PennState::SearchService::Client which
connects to Penn State's identity service and returns a set of person
records matching the search query.