Ingest new works from the API using actors
Updates IngestController and PublishNewWork to account for the changes
to depositors and actors. Now, when new works are published or migrated
from Scholarpshere 3, no new users will be created for each depositor.

Instead, the depositor is only an Actor, and if there are additional
permissions specified via the API, only those users who exist at Penn
State will have ACLs created for them on the work.