Allow ACLs to inherit other levels of access
Edit access should include read and discover access, and read access
should include discover access. Instead of using application logic to
effect this, additional ACLs are added that explicitly define these
inherited access levels.

Additionally, a `white_list` function is provided that serves to prevent
specific kinds of agents from being assigned using the `[]=` methods
available for each combination of agent and access level. This applies
to public and authorized group agents, so that they are not inadvertenly
removed when setting permissions for other agents.

Depositors are no longer checked when assigning permissions. This means
that the depositor will NOT appear as an edit user, and it's possible to
create an ACL for the depositor. However, these issues will be handled
within the application and the permissions module no longer has any
opinions about the depositor.