Default API token for development testing
This is purely for convenience. When we clean out the app, any tokens
that we had defined for the API get removed. This regenerates a token
that we can use to confiugure our other applications, such as the
Scholarsphere client and Scholarsphere 3 application. That way we don't
have to recreate their tokens each time, we can just use one for all of
them.

This is only in development, and for no other environment.