Move all stylesheet and image assets to webpacker
This refactors and reorganizes our webpacker code, moving in all our
other stylesheet and image assets.

The primary reason for this was because we were installing Bootstrap in
two places: the gem (for stylesheets), and the npm package (for
javascript). A recent jQuery security vulnerability exposed the flaw
with this setup: We could update jQuery (and consequently Boostrap) in
the npm package to address the vulnerability, but not the gem.

The solution is remove the asset pipeline altogether and the Bootstrap
gem, and only use webpacker. This necessitated some reorganizing of the
app/javascripts directory.

Co-authored-by: Charlie Morris <523381+cdmo@users.noreply.github.com>
Co-authored-by: Banu Hapeloglu Kutlu <bzk60@psu.edu>
Co-authored-by: Dann Bohn <djb44@psu.edu>