Add deposited_at attribute to resources
Resources that we migrate from Scholarsphere or potentially other
sources need to have their original deposit dates. Since these may be
older than the date the resources are created in the application, the
field must be a separate timestamp from the usual Rails created_at
field.

The deposited_at field is required for all Work, Collection, and
FileResource models. It defaults to the current time, but can be any
other valid timestamp.