Fix a potential Creator indexing bug
This fixes a slight, but very possible, bug with how and when Creator's aliases were stored.

A (WorkVersion|Collection)Creation object is what ties the resource to its creators, and stores the alias for that creator. Once this resource is published, the alias should _always_ stay the same, no matter if the Actor it links to renames themselves. To make life easier, this -Creation object will default its alias to the Actor's default_alias method. However, this defaulting was happening at "view time" not at "creation time" and therefore if it was left as null in the -Creation object,  the business rules above about never changing once the resource was published would fail.