Track views and downloads with ViewStatistic model (#390)
Adds a view statistic model with relationships to work versions,
collections, and file resources, in order to track views and downloads
of resources.

Requests from known bots are not counted using the Browser gem, and only
views for work versions are counted. Viewing a work increments the view
count for the latest published version only.