Migrate legacy identifiers for files
Adds a noid parameter for files coming over from Scholarsphere 3. The
noid is added as a legacy identifier to the FileResource. An additional
association is added to the file resource linking the old file set noid
in Scholarsphere 3 to the new file resource.

Noid creation is refactored out to a class method on LegacyIdentifier
for consistency.