Reload published work after successful save
When publishing a new work from the API, we should only reload it (and
thus get its newly minted UUID) if it has saved properly. Otherwise,
return the work object with its errors so that API can report those to
the caller.

This applies the same logic that is also in the CreateNewCollection
service.