Redis configuration is now a core service
Refactors the Redis configuration to be a core service. We no longer
check to see if the service is valid or active, it is assumed to be,
just like our database services. The configuration is persisted as a
hash within Rails' application configuration structure and can be
referenced from anywhere within application.

ActiveJob configuration is now kept separate from Redis and Sidekiq
tests are tied to this configuration setting instead.