Update to ViewComponent 2.x
Previous versions of this gem were under ActionView because at the time
the plan was that the gem would be added to Rails. That is now not the
case, and it will remain a separate gem outside of the Rails and
ActionView namespace.

This upgrades us to the new gem and changes our component classes
accordingly. The only major difference is that ViewComponent does not
support validation. But, since we didn't seem to be making much use of
that feature, removing it from our components doesn't affect the
outcomes of the tests.