Index draft versions of works
As a precursor to building the user's dashboard, we need to index draft
versions in Solr so we can search them. Because works are only indexed
if they are published, we can index additional work versions in draft
states without changing the behavior of search results in the main
application.