Linter fixes
Changes made to the original sass code in order to meet our linter
requirements. Most of these changes are minor, and when relevant, some
of the linters are disabled for specific files.

In a few cases, it would be helpful to revisit these later, such as the
linter errors for MergeableSelector and PlaceholderInExtend, but these
are not necessary to fix at this time.