Grant admins access via Pundit policies
Admins have a "can do anything" policy which our Pundit policies need to
reflect. In order to accomplish this, we had to touch each of the public
methods in our policies to ensure that admin could do all of them.
However, in order to make scoping a bit easier, we modified the Scope
superclass and implemented a new `#limit` method for the subclasses to
use instead of `#resolve`. This way, admins will always of have a scope
of `:all` by default and subclasses must implement `#limit` if they want
to change the scope of standard, non-admin users.