Display icons for work types
Selects a icon from the materials.io icon set based on the work type.
The icons are displayed for search results and dashboard results.

Works and work versions are indexed differently, and include the work
type identifying key as well as the display variation of the type.
Additionally, our icons are compiled into our webpack instead of being
loaded separately.