Deposit agreement version
Adds a deposit agreement version and timestamp to reflect the time when
a work is published. The work will be updated with the current deposit
agreement version when its first version is published. Subsequent
published versions will not change the timestamp so long as the deposit
agreement version remains unchanged.

If the deposit agreement version is changed, the work will NOT be
updated until a work version is publised.