Add work type to collection
While technically not a work, adding a work type definition to a
collection clears up a lot of other headaches. We can consistently
render thumbnail images for collections based off of that field, and we
can facet on collections now as well.

This avoids more complex logic in order to determine when we have a
collection, and also avoids a second field for faceting on collections.
Since there would be only two types: work or collection, it made sense
to combine a collection value into a work type (for now).