Searching and adding new creators
Adds functionality to the work form that searches for existing creators
(Actor objects) or people from Penn State's identity service. Creators
are added to the form via async callbacks. If the creator is not found,
a modal window prompts the user to create a new Actor record for the
creator.