Toggle admin privileges
Allows a user, who has admin rights available to them, to disable/enable
those rights via their profile page. This is useful if an administrator
wants to use the application like a standard user, or to verify access
controls to standard users are being observed properly.

This also updates the layout and formatting of the profile edit page to
make it more consistent with the rest of the forms in the application,
and ensure that the user does not have to enter an ORCiD id for their
profile.