Link dashboard work versions to their resources pages
Updates the dashboard search results to link work versions to their
resource pages. Includes a new NullWorkVersion object as a stand-in for
works that do have a published versions. This follows the "null object
pattern" and will help us in the future when dealing with works that
have no published version.