Update ORCiD validation
Removes the OrcidValidator class in favor of a simpler regex-based
verification that ensures the id will be stored in the database as a
16-digit string. The Orcid model is updated to handle verification and
translations between the different formats. While the model is not used
directly for validation, it can be used to validate data from other
sources, or present numeric ids in a more human-readable format.