Switch sidebar with main content
For accessibility reasons, it's preferred to have a screen reader
present the main content first, before it presents any sidebar content.
Because of layout reasons, this content was appearing *after* the
sidebar. Using Bootstrap's reordering classes, we can present the main
content fist in the page structure, but have it display second, after
the sidebar.