Operations in the user's dashboard
This tests for and corrects a couple of problems with the user's
dashboard:

* creating a new work version directs to the revised work form
* deleting a work version redirects back to the dashboard
* deleted versions are properly removed from the index and their parent
  works updated
* saving in the work form redirects back to the dashboard

Because most of the activities redirect back to the user's dashboard,
all Solr operations are performed *synchronously* so the changes can be
seen immediately.

At this point, we aren't making any provisions for asynchronous Solr
operations with regards to deleting work versions because there isn't a
clear use case at the moment.