Use GlobalID in PaperTrail's whodunnit
Uses the GlobalID to identify either a User or ExternalApp in the
whodunnit column of PaperTrail's audit information. For Rest API calls,
the token is used to set the current user to the associated application.
PaperTrail then uses this infomation to determine the global id of the
user or application that has made the changes to a work version.

Work history components are updated to display the access id or name in
the summary tab on the main resource page.