Subclass WorkVersion to index immidately in the dashboard
Dashboard::WorkVersion serves as an identical stand-in to WorkVersion
within the user's dashboard. This allows for synchronus updating of the
work version in Solr, so any changes will appear immediately to the
user. Outside of the dashboard, such as API calls, in the console, or
elsewhere, the updates are performed async via ActiveJob