Index work versions synchronously in the dashboard
When updating or creating work versions via the user's dashboard,
changes should be made to Solr immediately to avoid any lags in
displaying search results. This way, when a user creates or updates a
work version and then navigates back to the dashboard search page,
they'll see the changes instead of having to wait for the indexing job
to finish.