Merge Dashboard policies into main policies
This eliminates the Dashboard module of our polices and combines all of
the policy control to the main application policy classes. It also
includes a considerable amount of refactoring, and refining existing
policies to cover all the edge cases.

The policies now (officially) combine both ACL and non-ACL logic into one
place. For example, non-ACL based permissions such as access via
depositor or proxy ownership, as well as administrative access, are
combined with ACL-based permissions such as edit, read, and discover. As
a result, asking a policy for a type of access will interrogate all
those dimensions, both ACL and non-ACL.

Because of the varied nature of the permissions and because Solr drives
the search features, policy scopes are largely ignored. This is
reflected in the changes to controllers were scoping find operations are
either superfluous, or not accurate.