Display editable works in user's dashboard
The user's dashboard was not displaying work versions that the user had
edit access to via ACLs. This was because edit permissions were not
correctly indexed into the Solr record.

This fixes the above issue by adding both edit and read permissions to
the PermissionsSchema, which in turn indexes these permissions in works,
work versions, and collections.

Additionally, another bug was discovered when trying to actually edit
these resources. The ACLs cannot be matched against UserDecorator
objects. The ApplicationPolicy was updated to cast any decorated user
records to their ActiveRecord source.