Render markdown partials with custom markdown handler
This registers an ActionView template handler for rendering partials
written in mardown. Redcarpet is used to convert the markdown into HTML,
and the rest relies on Rails' internal templating architecture.

The MarkdownController defines a single action to render the page.
Specific pages are defined in routes, linking the name of the route to
the corrsponding markdown file.