Contact form
Creates an email form identical to the one used in Scholarsphere 3 to
report issues with the application. Emails that are send create tickets
in ServiceNow, as well as a confirmation email sent to the user who has
raised this issue.

Confirmation emails only get sent to Penn State email addresses.