Update search results display
This changes the both the public and dashboard search so that all
results are displayed the same way. The deposited date for works and
collections is used, there are no edit or delete links used, history and
work type is no longed displayed, and instead the first three creators
are shown.