Rake task for async reindexing of works
Updates our existing rake task for reindexing works to operate async,
putting a job on the queue for each work. This requires refactoring our
update_index methods so that committing can be specified at each point
along the way.

For each *work* that is indexed, commits are not performed until the
the job for the last work is put on the queue.