Required elements for publishing
License and deposit agreement are now marked required, which will
inform users using HTML5 feature and will not longer display the error
messages if they are navigating to the page for the first time.

We have also updated the visibility label to read _Access_, making it
consistent throughout the application.