Disallow Works to be shown if they are not published
When rendering a Work, the resource page requires/assumes that there
will be at least one published version. This was causing trouble when a
user _with edit access_ requested the resource page for a draft work,
which caused a 500. This commit tweaks the policy to raise a pundit
error when requesting the resource page of draft work.