Collection form for managing collections
Duplicates the same user interface and workflow that's used to create
and edit works. Included in this is a MASSIVE refactoring of all the
controller and view code that was in the original work form. In order to
reduce the duplication of controller and view code, the previous
WorkForm module is renamed to a Form module, and incorporates collection
controllers as well.

Controllers for contributors are identical for both work versions and
collections, while details are more similar and the view layer.

Routing is now slightly different, eschewing the /dashboard/work-form
path in favor of /dashboard/form/work_version or
/dashboard/form/collection.

Duplicates the work form controllers and views in order to build a form
for creating new collections.