Use DisplayRightsComponent for licenses
Translates a give license id to its corresponding label, and makes the
link to the label's URI. The component does not render if the id is
missing or invalid.

There was also a bug with license labels. A nil or bogus value would
cause an error. This is now corrected and a nil label is returned.

Co-authored-by: Seth Erickson <sr.erickson@gmail.com>