Respond with json for errors in non-html formats
Custom error responses need to account for the format of the original
request. If a user requests a resource in a format other than html, such
as xml, rss, or other, the error response will attempt the respond using
the requested format. This was causing missing template errors because
the application was trying to honor the request with the requested
format.

Since we don't want to attempt to match every kind of non-html response,
we'll just respond with json for any non-html request that generates an
error.