Support deleting collections
Adds a delete button to the collection resource page which deletes the
collection with a confirmation dialog. Collections are now deleted from
Solr synchronously. We may want to revisit this later and adopt the
pattern similar to updating indexes via a source method that can be
either async or sync, but that is out of scope at this moment.