Remove references to file editor controller
Previously, filename_editor_controller.js enable inline editing of
filenames, but that feature has been disabled for now. However, the
stimulus controller was still being loaded in the form, which was
causing errors in the console. Although this didn't affect the function
of the form, it's better to not have the errors popping up.

References to the js controller have been removed, but the controller is
still there in case we want to reenable these features in the future.