Index resources only when completing or exiting the form
When editing work versions or collections, it's only necessary to index
the resources in Solr when completing the last step of the form, or when
exiting in the middle. For work versions, this means that we'll only
index the work in Solr if publishing is successful, or if the user
saves and exits the form. Likewise, with collections, they will only be
indexed after the final "finish" step when members are added, or if the
user saves and exits the form.

This will improve performance between steps, since no sync. indexing will
be done, and will hopefully avoid any issues with multiple index jobs
overwriting one another.

Indexing will always be sync. or won't happen at all.