Catch user errors in the editors form (#800)
If the user enters a full name, with spaces, this raises an
URI::InvalidURIError. We'll catch this and return a friendlier error
message to the user. Otherwise, we'll let the exception proceed like
it's supposed to. This has the downside of a less-than-ideal user
experience, but it means we'll be able to see unexpected errors from our
integration with Penn State's identity service.