Add upload endpoint to API
Returns a presigned url to a client so that they may upload a file into
our S3 instance. The endpoint requires an extension, such as pdf, or
txt, so that it can construct a path for the file in the bucket.

Filenames are unique UUIDs, matching the current behavior in the UI via
the Shrine/Uppy process.