Remodel creators as authors for works and collections
This is the first commit of a multi-commit process to remodel creators
as authors, removing the requirement that every creator also be tied to
a unique actor. This will support the ability to add creators to a work
or collection without requiring a unique identifier such as a Penn State
access id or Orcid.

A new, polymorphic Authorship table is added to replace the existing
WorkVersionCreation and CollectionCreation.  An authorship record has an
optional link to an actor so that if the creator does have a unique
identifier, we can link each of their authorships to a single,
disambiguated person. If the creator has no identifier, then they
have no corresponding Actor record.

While the database schema is altered, no data should be modified or
destroyed with this commit. Migrating existing data into the new schema
will be in a future commit and will be an intensive process.