Update DoiService to accept Collections too
This was a big refactor, where ultimately I made two internal strategy classes, one for handling Works and Versions, and another for handling Collections.

I also un-DRY'd some of this code, which RuboCop did complain about (and surely CodeClimate will too). The reason I did this was because, when refactoring, the code was so DRY that it was really, really hard to figure out what it was actually doing. Essentially we've got a four-state truth table, which does four similar but slightly different things. I decided to just write this all out rather than DRY it up because I think it's much eaiser to follow.