I#313 collection dois (#838)
* Refactor DataCite Metadata adapter for WorkVerisons and Collections

Previously we had a class `DataCite::Metadata` that took a WorkVersion and turned it into the metadata schema that DataCite requires to publish a DOI.

With the advent of minting DOIs for Collections, I needed to write a corresponding metadata adapter to translate a Collection's metdata into datacite format.

* Update DoiService to accept Collections too

This was a big refactor, where ultimately I made two internal strategy classes, one for handling Works and Versions, and another for handling Collections.

I also un-DRY'd some of this code, which RuboCop did complain about (and surely CodeClimate will too). The reason I did this was because, when refactoring, the code was so DRY that it was really, really hard to figure out what it was actually doing. Essentially we've got a four-state truth table, which does four similar but slightly different things. I decided to just write this all out rather than DRY it up because I think it's much eaiser to follow.

* Refactor DoiService

* DataCite metadata mappers support EDTF publication dates

* Refactor Datcite metadata error classes out of base class