Update creators' positions when null (#834)
* Update creators' positions when null

Due to a migration oversight, migrated works did not have any position
indicators. Additionally, there was a bug that was not copying over a
creator's positioning indicators to new versions. The bug has since been
fixed, but there remained many creator records will null positions.

This is a service class, designed to run in the console only, and add
any missing positions. Any first version without positions has them
added according to the order of the primary key. This represents the
order in which they were added to the database. Any subsequent versions
that don't have positions have them copied over from their previous
versions.

The process is idempotent and can be re-run multiple times and never
affect creators that have position values. If a version is encountered
that somehow has a mix of creators with and without null positions, it
aborts.

* Trap and log errors in CreatorOrderFix

Co-authored-by: Ryan Schenk <rschenk@gmail.com>