Migrate Authorships (#854)
Moves existing data into the new Authorship data model. This includes moving WorkVersionCreation records, and CollectionCreation records, to Authorship records.

In the process of working through this, we made a couple of changes to the way versions are tracked in PaperTrail. Originally, we did *not* register create events when copying creators to new work versions. This was to make the change history look accurate, otherwise a new version would register the _creation_ of a bunch of new creators even through nothing had changed. In the course of migrating these creators to the new authorship model, it became clear that having those original creation events would have made life a lot easier. To that end, we are now capturing all PaperTrail events with creators, but flagging those that we do not wish to show in the version history.

Another side-effect of this data migration is that several works had duplicate creators removed. This problem went all the way back to Scholarsphere 3 metadata where the same creator was listed twice. In the migration to Scholarsphere 4, these duplicate creators were retained, but were now linked to the same actor record. During the authorship migration, these duplicate creators were removed.