Add reload_on_index option to work versions (#897)
When indexing works and their versions, the versions can sometimes be
out-of-sync with the database. Currently, this only happens during the
publish phase if the visibility is set to Penn State. Just after the new
version is published:

    @resource != @resource.work.versions[0]

Where @resource is the version to be published, and the work only has
one version. Because the _work_ is submitted to the indexer, the
out-of-date work version gets indexed instead of the current @resource
work version.

We don't know why this happens, but the fix for now is to signal the
WorkIndexer to reload the work if the version is being published.