Orcid id is not present for unidentified creators (#900)
Since a creator (Authorship) may or may not have an associated actor,
the Orcid will only be present if the actor is. For now, we can
interrogate the authorship about the presence or absence of an Orcid.

In the future, however, we may want to adopt an null object pattern if
we keep having to check for the presence of an actor or not.