BuildNewActor service
To support API integrations, we need to be able to build new or existing
Actor objects given an identifier. This way, a client can provide either
a Penn State access id or an Orcid id, and we'll be able to create the
required Actor without any additional information.

This pattern assumes the client knows the id and that it is valid;
otherwise, an error will (ultimately) be returned to them.