Change PublishNewWork to WorkPublisher service
The original PublishNewWork service has completely different
requirements now that we no longer need to support migration from
Scholarsphere 3.

To better reflect the link between the IngestController, the service is
now renamed to API::V1::WorkPublisher so that it aligns with the API's
version namespace.

The notable changes to the new service include:

* accepting creators without identifiers
* actor records are only created for identified creators
* no default values; everything must be explicitly declared
* dropping support for legacy identifiers (via API)
* works are only created when published; no draft works
* all validations should align with the UI
* creators can be added using Penn State access ids or Orcids