Do not raise errors for invalid visibilities
When we raise an error for an invalid visibility, it prevents us from
properly reporting the error via ActiveModel::Errors because we're not
able to return the original object.

If a visibility level is not supported, just do nothing: do not change
any ACLs. Visibility is validated at the work version, and the error
will be reported to the user via `errors`.