Require a md5 checksum for presigned urls
Our client gem will automatically calculate checksums for files that we
upload into S3, but our API must include the checksum when requesting
the presigned URL. Instead of making the parameter optional, which would
affect the behavior of the client, the parameter should just be
required. This is better overall for data preservation and ensuring
files make it into S3 correctly.